Added files via upload
This exe and txt file is just needs to be placed in the same directory (no install) and is directly runnable program for Windows computers with .net already installed.  it will let you select a Drive/Directory and will scan for files with known Crytolocker/Wall files or known file types that are not standard (Based on File Signature).  the TXT File can be updates with more "good" or "bad" file types based on Signatures, the file names can be exported as .CSV files.   -- Use at your own risk, we will not be responsible for any damages caused by this program.